requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Documents
pymupdf>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.13

# Retrieval
chromadb>=0.5.0
sentence-transformers>=3.0.0

python-multipart>=0.0.9

# Tests
pytest>=8.3.0
